predictionAPI/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
pydantic==2.10.3
pandas==2.3.0
numpy==2.1.3
scikit-learn==1.7.0
joblib==1.5.1
google-generativeai==0.8.5
textblob==0.19.0
xgboost==3.0.2
tensorflow==2.19.0
keras==3.10.0
h5py==3.14.0
typing-extensions==4.12.2
python-dotenv==1.1.1
